+++ source/auth/auth_ntlmssp.c
@@ -114,13 +114,15 @@ static NTSTATUS auth_ntlmssp_check_password(struct ntlmssp_state *ntlmssp_state,
 		return nt_status;
 	}
 	if (auth_ntlmssp_state->server_info->user_session_key.length) {
-		DEBUG(10, ("Got NT session key of length %u\n", auth_ntlmssp_state->server_info->user_session_key.length));
+		DEBUG(10, ("Got NT session key of length %u\n",
+			(unsigned int)auth_ntlmssp_state->server_info->user_session_key.length));
 		*user_session_key = data_blob_talloc(auth_ntlmssp_state->mem_ctx, 
 						   auth_ntlmssp_state->server_info->user_session_key.data,
 						   auth_ntlmssp_state->server_info->user_session_key.length);
 	}
 	if (auth_ntlmssp_state->server_info->lm_session_key.length) {
-		DEBUG(10, ("Got LM session key of length %u\n", auth_ntlmssp_state->server_info->lm_session_key.length));
+		DEBUG(10, ("Got LM session key of length %u\n",
+			(unsigned int)auth_ntlmssp_state->server_info->lm_session_key.length));
 		*lm_session_key = data_blob_talloc(auth_ntlmssp_state->mem_ctx, 
 						   auth_ntlmssp_state->server_info->lm_session_key.data,
 						   auth_ntlmssp_state->server_info->lm_session_key.length);
